# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Users API",
    "description": "This is docs for the devices and management api",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "tags": [
    {
      "name": "USERS",
      "description": "Users documentation"
    },
    {
      "name": "Bus",
      "description": "Bus API documentation"
    },
    {
      "name": "Bookings",
      "description": "Booking documentation"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "summary": "Login user",
        "description": "Authenticates a user",
        "tags": ["USERS"],
        "requestBody": {
          "$ref": "#/components/requestBodies/Login"
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "$ref": "#/components/responses/LoginSuccess"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "User not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/register": {
      "post": {
        "summary": "Register user",
        "description": "Creates a new user",
        "tags": ["USERS"],
        "requestBody": {
          "$ref": "#/components/requestBodies/Register"
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "$ref": "#/components/responses/RegisterSuccess"
          },
          "400": {
            "description": "Bad request"
          },
          "409": {
            "description": "User already exists"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/bus/search": {
      "post": {
        "tags": ["Bus"],
        "summary": "Search Bus with summary of seats",
        "requestBody": {
          "$ref": "#/components/requestBodies/Search"
        },
        "responses": {
          "201": {
            "description": "Fetched bus's seats on specific date successfully ",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/bus/seats": {
      "post": {
        "tags": ["Bus"],
        "summary": "Search detailed seats overview",
        "requestBody": {
          "$ref": "#/components/requestBodies/Search"
        },
        "responses": {
          "201": {
            "description": "Fetched detailed seats overview",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/bus": {
      "get": {
        "tags": ["Bus"],
        "summary": "List all buses",
        "responses": {
          "201": {
            "description": "Listed all the buses successfully",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/bookings/book": {
      "post": {
        "tags": ["Bookings"],
        "summary": "Book seat any day",
        "requestBody": {
          "$ref": "#/components/requestBodies/Booking"
        },
        "responses": {
          "201": {
            "description": "Booked seat successfully",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/bookings/{id}": {
      "get": {
        "tags": ["Bookings"],
        "summary": "User-History",
        "parameters": [
          {
            "$ref": "#/components/parameters/UserIdParam"
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully fetched user-history",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
    "/bookings/cancel": {
      "post": {
        "tags": ["Bookings"],
        "summary": "Cancel ticket",
        "requestBody": {
          "$ref": "#/components/requestBodies/Cancel"
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {}
            }
          }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "UserIdParam": {
        "name": "id",
        "in": "path",
        "description": "Provide **Particular ID**\n",
        "required": true,
        "schema": {
          "type": "integer",
          "example": 241
        }
      }
    },
    "requestBodies": {
      "Booking": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Booking"
            },
            "example": {
              "seat_id": 11,
              "bus_id": 4,
              "date": "2024-11-19",
              "user_id": 241
            }
          }
        }
      },
      "Cancel": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Cancel"
            },
            "example": {
              "schedule_id": 21
            }
          }
        }
      },
      "Register": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RegisterRequest"
            },
            "examples": {
              "RegisterCorrect": {
                "$ref": "#/components/examples/RegisterCorrect"
              },
              "RegisterIncorrect": {
                "$ref": "#/components/examples/RegisterIncorrect"
              }
            }
          }
        }
      },
      "Login": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LoginRequest"
            },
            "examples": {
              "LoginCorrect": {
                "$ref": "#/components/examples/LoginCorrect"
              },
              "LoginIncorrect": {
                "$ref": "#/components/examples/LoginIncorrect"
              }
            }
          }
        }
      },
      "Search": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Search"
            },
            "example": {
              "bus_id": 4,
              "date": "2024-11-19"
            }
          }
        }
      }
    },
    "responses": {
      "RegisterSuccess": {
        "description": "User registered successfully",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/RegisterResponse"
            }
          }
        }
      },
      "LoginSuccess": {
        "description": "User logged in successfully",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/LoginResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "Cancel": {
        "type": "object",
        "properties": {
          "schedule_id": {
            "type": "integer"
          }
        }
      },
      "Booking": {
        "type": "object",
        "properties": {
          "seat_id": {
            "type": "integer"
          },
          "bus_id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "date": {
            "type": "string"
          }
        }
      },
      "Search": {
        "type": "object",
        "properties": {
          "bus_id": {
            "type": "integer"
          },
          "date": {
            "type": "string"
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "format": "email",
            "pattern": "^[\\w-\\.]+@([\\w-]+\\.)+[\\w-]{2,4}$"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          },
          "firstname": {
            "type": "string"
          },
          "lastname": {
            "type": "string"
          },
          "mobile": {
            "type": "string",
            "pattern": "^\\+\\d{1,3}\\s?\\d{1,14}$"
          },
          "address": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password",
          "firstname",
          "lastname",
          "mobile",
          "address"
        ]
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "enum": ["User registered successfully"]
          },
          "user": {
            "type": "object"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "enum": ["User logged in successfully"]
          },
          "token": {
            "type": "string"
          },
          "user": {
            "type": "object"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "format": "email",
            "pattern": "^[\\w-\\.]+@([\\w-]+\\.)+[\\w-]{2,4}$"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": ["username", "password"]
      }
    },
    "examples": {
      "RegisterCorrect": {
        "value": {
          "username": "selva@gmail.com",
          "password": "12345678",
          "firstname": "Selva",
          "lastname": "Kumar",
          "mobile": 9876543210,
          "address": "123 Main St"
        }
      },
      "RegisterIncorrect": {
        "value": {
          "username": "selva@gmail.com",
          "password": "12345678",
          "firstname": "Selva",
          "lastname": "Kumar",
          "mobile": 9876543210,
          "address": "123 Main St"
        }
      },
      "LoginCorrect": {
        "value": {
          "username": "naveen@gmail.com",
          "password": "12345678"
        }
      },
      "LoginIncorrect": {
        "value": {
          "username": "naveen@gmail.com",
          "password": "123"
        }
      }
    }
  }
}